requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
gast==0.3.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14.1
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
joblib==0.16.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.0
numpy==1.26.1
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
tensorboard==2.17.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.17.0
tensorflow-estimator==2.15.0
termcolor==2.4.0
threadpoolctl==3.5.0
urllib3==2.0.7
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.1